Features/47 shallow parsing (#78)
When initially scanning a project it is useless to scan the contents of methods - as the variables in those methods don't matter externally. When a user starts editing a file and actually types anything, the file they are editing will get refreshed and populate the variables for that file - which is all they really need.

Oh. Except for superclasses, etc. That's kind of a bummer, but I'm not sure how well that worked before. Certainly no tests fail...